feat: Allow use of non-CA certs in TLS client authentication (#4)
Motivation
In some cases, we would like to enable an "intra-service" TLS client auth whereby all peers have the same key/cert pair but not necessarily the corresponding CA cert. This does not work with the default Java X509 TrustManager behaviour unless the certificate's issuer and subject match (which they generally won't if the cert is not a CA). The problem is that the subjects of the certs returned from TrustManager#getAcceptedIssuers() are used in the client certificate request handshake message, and the client will fail the handshake if it receives only CA names that do not match its cert's issuer.

Modifications
Wrap the TrustManager in an implementation which delegates all methods but will return an empty list from getAcceptedIssuers() if any of the configured trust certs are non-CA. When an empty CertRequest is received the client will attempt to use any cert for authentication. This is still secure since the usual verification is done (and passes) on the server-side.

Result
TLS client auth works when the server has the client key's cert but not CA cert.

Co-authored-by: nickhill <nickhill@us.ibm.com>