feat: Support configuration of service class name aliases
Motivation

Litelinks performs validation that the thrift service class name matches when a client discovers a given service and when servers join an existing cluster. This is to ensure that the service interface matches. However, it also means there is no way to change the name of the service interface definition and upgrade existing clusters in-place because the new servers will refuse to join the cluster.

Modifications

- Introduce a new environment variable that can be used to define aliases for the service class names, so that the old name will be interpreted as the new one. This can then be added to existing clients and new server builds.
- Rename the ThriftConnProp class to more appropriate "ServiceProperties"
- New unit test which makes use of a new dummy thrift service definition

Result

It is possible to change litelinks service class names and upgrade existing clusters in a non-disruptive manner.