feat: Add "abort request" option for custom load balancers (#10)
feat: Add "abort request" option for custom load balancers

Motivation

Litelinks supports custom client load balancing policies via the LoadBalancer and LoadBalancingPolicy interfaces. For a given getNext() call, an LB implementation may return null to reject all of the "offered" service instances. This is used in some cases to short-circuit the RPC request when the chosen service instance corresponds to the local process, so that a more efficient direct invocation can be made.

However, litelinks interprets the null return as "no suitable instances" from the "active" list and before returning from the client method call will first attempt sending to an instance on the "failing" list if it is non-empty.

Modifications

To allow the short-circuiting to work in cases that there are failing instances, we add an ABORT_REQUEST constant to the LoadBalancer interface which can be returned as an alternative to null, but will instruct model-mesh to fail the request immediately rather than attempting to fall back to any failing instances.

Also extended unit test to cover the new feature and simplified some formatting in related classes.

Result

Custom load balancers have a reliable way to abort an RPC.